fix bug: https://bugs.gentoo.org/908254
add systemd use flag to ebuild, so the dependency
on systemd is explict.
The package did silently depend on systemd in case it was installed on
the build machine.

About the patch: cherry pick merged from:
Author:     Sam James <sam@gentoo.org>
meson: add option for systemd

Without this, systemd will be used if installed on the system automagically,
which is a problem if the built e.g. Xwayland is going to be used on a non-systemd
machine.

Bug: https://bugs.gentoo.org/908254
Signed-off-by: Sam James <sam@gentoo.org>
[Cherry pick merged by alexander-n8hgeg5e <alexander-n8hgeg5e@alex-cloud.de>]